Scale z-coordinate for screen coordinates
Screen coordinates are used in the calculation of baryonic coordinates,
but the calculation of the z-coordinate is done in a way that breaks the
invariant of straight lines remaining straight.

This fix divides the z-coordinate by the same as the x and y coordinates
are divided.

z' = (z-1)/z

This projects z = [1..+inf] onto z' = [0..1]